refactor(engine): Use SendCorrelateCommandSideEffectProducer
Move inner class to top level. Reuse class in second call site.
