+++ engine/src/main/java/io/camunda/zeebe/engine/processing/message/MessageCorrelator.java
@@ -15,9 +15,7 @@ import io.camunda.zeebe.engine.state.message.StoredMessage;
 import io.camunda.zeebe.protocol.impl.record.value.message.MessageSubscriptionRecord;
 import io.camunda.zeebe.protocol.record.intent.MessageSubscriptionIntent;
 import io.camunda.zeebe.scheduler.clock.ActorClock;
-import io.camunda.zeebe.util.buffer.BufferUtil;
 import java.util.function.Consumer;
-import org.agrona.DirectBuffer;
 import org.agrona.collections.MutableBoolean;
 
 public final class MessageCorrelator {
@@ -82,43 +80,5 @@ public final class MessageCorrelator {
     }
 
     return correlateMessage;
-  }
-
-  private static final class SendCorrelateCommandSideEffectProducer implements SideEffectProducer {
-
-    private final SubscriptionCommandSender commandSender;
-    private final long processInstanceKey;
-    private final long elementInstanceKey;
-    private final DirectBuffer bpmnProcessIdBuffer;
-    private final DirectBuffer messageNameBuffer;
-    private final long messageKey;
-    private final DirectBuffer variablesBuffer;
-    private final DirectBuffer correlationKeyBuffer;
-
-    private SendCorrelateCommandSideEffectProducer(
-        final SubscriptionCommandSender commandSender,
-        final MessageSubscriptionRecord subscriptionRecord) {
-      this.commandSender = commandSender;
-
-      processInstanceKey = subscriptionRecord.getProcessInstanceKey();
-      elementInstanceKey = subscriptionRecord.getElementInstanceKey();
-      bpmnProcessIdBuffer = BufferUtil.cloneBuffer(subscriptionRecord.getBpmnProcessIdBuffer());
-      messageNameBuffer = BufferUtil.cloneBuffer(subscriptionRecord.getMessageNameBuffer());
-      messageKey = subscriptionRecord.getMessageKey();
-      variablesBuffer = BufferUtil.cloneBuffer(subscriptionRecord.getVariablesBuffer());
-      correlationKeyBuffer = BufferUtil.cloneBuffer(subscriptionRecord.getCorrelationKeyBuffer());
-    }
-
-    @Override
-    public boolean flush() {
-      return commandSender.correlateProcessMessageSubscription(
-          processInstanceKey,
-          elementInstanceKey,
-          bpmnProcessIdBuffer,
-          messageNameBuffer,
-          messageKey,
-          variablesBuffer,
-          correlationKeyBuffer);
-    }
   }
 }
